Add score models (for... not-best scores)
